+++ product/src/main/java/takeoff/logistics_service/msa/product/product/application/client/UserClient.java
@@ -1,4 +1,4 @@
-package takeoff.logistics_service.msa.product.product.application.service;
+package takeoff.logistics_service.msa.product.product.application.client;
 
 import takeoff.logistics_service.msa.product.product.application.dto.response.GetUserResponseDto;
